refactor: complete graphics engine overhaul
For anyone tracking, this fundamentally rewrites the graphics engine from the ground up. To be short, the old paradigm was based on groups of line segments drawn into infinite space. The fallacy was that I would always have room to optimize culling that set of edges before drawing with a traditional left-to-right complete raycast scan.

This refactors the engine to be closer to Doom (but without BSP, for now), in that levels are made up of convex polygons, and all textures are mapped to interior walls, even in supposed "outdoor" environments. Rather than deal with BSP right away, I just added convex polygon editing to the editor and dropped the old style of graphics.

Note that the city of Bloomington is a shadow of its former self, mainly because it was only ever meant as a test city, and it will be axed soon anyway. But, in addition to the gigantic framerate boost, this also gets a shiny new palace area prototype, though there is nothing to do in it, and Bloomington remains the only place with interactive elements, for now.

That should all change soon with the new freedom granted by this performance boost.