feat: dynamic texture loader
This is a behind-the-scenes change with no immediate impact on the public build, but it paves the way for more content.

Previously, all textures in the game lived in the main cartridge, but I was nearly out of room. I now store all textures in separate PNG files in my workspace, which are imported by the build system and stuffed into a new textures.p8 cart, but stored in sequential data order all throughout the cart instead of as 16x16 sprites the way pico-8 normally would. Then, at runtime, the data for a single texture can be loaded on its own and copied into the correct format into the next available space. While there is still a 47-texture limit at runtime, this allows me to store up to 134 total textures in the textures.p8 cart.